requirements.txt
flask==2.3.3
flask-mysqldb
werkzeug==2.3.7
python-dotenv
